Connect url-to-pdf api to frontend

+++ frontend/src/components/custom/book_page/tests/QuizSolver.tsx
@@ -70,7 +70,7 @@ const QuizSolver: React.FC = () => {
         query: mutation,
         variables: {
           id: parseInt(testId!),
-          lastResult: parseFloat(resultPercentage),
+          lastResult: parseInt(resultPercentage),
         },
       }, {
         headers: { 'Content-Type': 'application/json' },
